Moved fonts/stickers/etc to assets
