+++ Brushshe/assets/palettes/4bit_palette.hex
