+++ Brushshe/assets/palettes/default_palette.hex
