+++ Brushshe/assets/palettes/vintage_palette.hex
