+++ Brushshe/assets/stickers/face2.png
