+++ Brushshe/assets/stickers/heart.png
