+++ Brushshe/assets/stickers/strawberry.png
