+++ Brushshe/core/bpoint.py
@@ -1,0 +1,2 @@
+class Bpoint():
+
